secure-mission-tasking-portal/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
marshmallow==3.21.3
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pytest==8.2.2
